Incorporated the Authorized Agreement and Conditions
The Erebus theme is subject to the Creative Commons 3.0 License. This means there's a clear and open framework for sharing, reusing, and adapting content by others. In doing so, a collaborative and creative environment is fostered, allowing fellow creators to engage with and build upon the work while ensuring proper attribution. This license also ensures the responsible and transparent use of other creators' work, giving credit where it's due to contribute to a culture of openness and shared innovation.